+++ django/songdecks/songdecks/views/commanders.py
@@ -1,5 +1,6 @@
 from rest_framework.decorators import api_view
-from django.http import JsonResponse
+from rest_framework.response import Response
+from rest_framework import status
 from songdecks.serializers import (CommanderSerializer)
 from songdecks.models import (Faction, Commander)
 from songdecks.views.helpers import upload_file_to_s3
@@ -13,13 +14,21 @@ def get_commanders(request):
     try:
         commanders = Commander.objects.all()
         serializer = CommanderSerializer(commanders, many=True)
-        return JsonResponse({"success": True, "response": serializer.data})
+        return Response(serializer.data, status=status.HTTP_200_OK)
     except Exception as e:
-        return JsonResponse({"success": False, "response": str(e)})
+        return Response(
+            {"detail": str(e)},
+            status=status.HTTP_500_INTERNAL_SERVER_ERROR
+        )
 
 @api_view(['POST'])
 def add_edit_commander(request, commander_id=None):
     try:
+        if request.user.profile.moderator == False:
+            return Response(
+                {"detail": "You do not have permission to delete commanders."},
+                status=status.HTTP_403_FORBIDDEN
+            )
         info = {
             'name': request.data.get('name', None),
             'img_url': request.data.get('img_url', None),
@@ -27,17 +36,26 @@ def add_edit_commander(request, commander_id=None):
         }
         for key in info:
             if info[key] is None:
-                return JsonResponse({"success": False, "response": f"Missing {key}."})
+                return Response(
+                    {"detail": f"Missing {key}"},
+                    status=status.HTTP_400_BAD_REQUEST
+                )
             
         img_file = request.data.get('img_file', None)
         if img_file:
             is_success, error_msg = upload_file_to_s3(img_file, AWS_S3_BUCKET_NAME, info['img_url'])
             if not is_success:
-                return JsonResponse({"success": False, "response": error_msg})
+                return Response(
+                    {"detail": error_msg},
+                    status=status.HTTP_500_INTERNAL_SERVER_ERROR
+                )
 
         faction_search = Faction.objects.filter(id=info['faction_id'])
         if faction_search.count() == 0:
-            return JsonResponse({"success": False, "response": "Faction not found."})
+            return Response(
+                {"detail": "Faction not found."},
+                status=status.HTTP_400_BAD_REQUEST
+            )
         if commander_id is None:
             commander = Commander.objects.create(
                 name=info['name'],
@@ -47,7 +65,10 @@ def add_edit_commander(request, commander_id=None):
         else:
             commander = Commander.objects.filter(id=commander_id)
             if commander.count() == 0:
-                return JsonResponse({"success": False, "response": "Commander not found."})
+                return Response(
+                    {"detail": "Commander not found."},
+                    status=status.HTTP_404_NOT_FOUND
+                )
             commander = commander.first()
             commander.name = info['name']
             commander.img_url = info['img_url']
@@ -55,19 +76,36 @@ def add_edit_commander(request, commander_id=None):
             commander.save()
 
         new_commander = Commander.objects.filter(id=commander.id).first()
-        message = f"Successfully created: {new_commander.name}" if commander_id is None else f"Successfully edited: {new_commander.name}"
-        return JsonResponse({"success": True, "response": message, "commander": CommanderSerializer(new_commander).data})
+        serializer = CommanderSerializer(new_commander)
+        return Response(serializer.data, status=status.HTTP_200_OK)
     except Exception as e:
-        return JsonResponse({"success": False, "response": str(e)})
+        return Response(
+            {"detail": str(e)},
+            status=status.HTTP_500_INTERNAL_SERVER_ERROR
+        )
     
-@api_view(['GET'])
+@api_view(['DELETE'])
 def delete_commander(request, commander_id):
     try:
+        if request.user.profile.moderator == False:
+            return Response(
+                {"detail": "You do not have permission to delete commanders."},
+                status=status.HTTP_403_FORBIDDEN
+            )
         commander_search = Commander.objects.filter(id=commander_id)
         if commander_search.count() == 0:
-            return JsonResponse({"success": False, "response": "Commander not found."})
+            return Response(
+                {"detail": "Commander not found."},
+                status=status.HTTP_404_NOT_FOUND
+            )
         commander = commander_search.first()
         commander.delete()
-        return JsonResponse({"success": True, "response": "Successfully deleted commander."})
+        return Response(
+            {"detail": f"Successfully deleted: {commander.name}"},
+            status=status.HTTP_200_OK
+        )
     except Exception as e:
-        return JsonResponse({"success": False, "response": str(e)})
+        return Response(
+            {"detail": str(e)},
+            status=status.HTTP_500_INTERNAL_SERVER_ERROR
+        )
